src/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch

# Embeddings
sentence-transformers==3.0.0

# ML clásico
scikit-learn>=1.4.0
joblib>=1.3.0

# Datos
pandas>=2.0.0
numpy>=1.26.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
email-validator>=2.2.0

# Autenticación JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Base de datos SQL Server
pyodbc>=5.1.0
psycopg2-binary>=2.9.9

# Reglas
pyyaml>=6.0
simpleeval>=0.9.13

# Utils
python-dotenv>=1.0.0
structlog>=24.0.0

# Dev / Test
pytest>=8.0.0
httpx>=0.27.0
jupyter>=1.0.0

hf_xet

psycopg2-binary==2.9.11
